requirements.txt
azure-identity
azure-keyvault-secrets
azure-functions
